Pass the item value to the selectionAdded callback (after the element parameter).
When adding an item, it can be helpful to know the item's value, beyond the basic visual element that may only contain the display name (label).  A practical use case would be to store this item value as a data attribute on the element itself so it can be easily accessed later.

Without this patch and saving the value in a data attribute, determining the value for an element would require matching its position within the widget to the value sequence in the comma-delimited list that autoSuggest maintains in its hidden input.  Unfortunately, even that is impossible if we wish to determine the value of elements added during preFill, since the hidden input (with comma-delimited values) does not yet exist in the DOM until after preFill completes.  For such a case, explicit passing of the item value is essential.