Optimizes internal handling of user input and server output (XSS protection).
In general, user inputs have to be escaped before using jQuery's code-injection
features like $(..).html(maliciousUserInput). That is guaranteed now. Test
cases are attached!

Additionally, this will add support for custom html renderer. A jQuery object
which is defined via configuration key "selectedItemProp" will be appended
directly to the result or selection list. Tests cases are attached.

Because of both changes (XSS protection and custom HTML renderers), the
highlighter had to be changed to work with html escaped content. Tests cases
are attached.

In preparation for an internal rewrite of data storage I had introduced the
local "Selections" holder. Instead of trusting on a hidden field with comma
separated strings, this is a real domain object. It supports add(value),
remove(value), exists(value), getAll() and clear(). Nearly all usages of
the hidden field were replaced in favor of "Selections". For backwards
compatibility, the hidden field will be synchronized on every modification.

Note: The XSS execution issues will work with the current versions.

Following test cases will be appended
- XSS: Type a malicious code fragment which must not be executed
  (new token, no result list).
- XSS: Type a part of malicious code fragment which is part of a result
  list. Neither the result list's nor the selection list's code must be
  executed.
- Extended example/test using a custom formatter with "formatList" (basic).
- Extended example/test using a custom formatter and prefilling "preFill".
  The prefilled data should generate the selection correctly (basic).
- Extended example/test using a custom formatter, prefilling and a
  custom selection renderer.
- A combination of all tests: custom result list item renderer, custom
  selection list item renderer and XSS check.