Migration to grunt, fixes regression test after CoffeeScript transforming, adds SASS, adds Mins.
# 1: I'm introducing the usage of `grunt`.
This will replace the old `make.sh` and several other manual required steps. To verify a change, just type `grunt test`. To make a new build into destination, just type `grunt build`. The default and recommended way is only `grunt` which all compile all, test all and make all required build artifacts at the destination.

Based on a common `grunt` setup this projects have a set of several additional tasks: coffee, coffeelint, cssmin, sass and a misc (utils). Credits by https://github.com/Takazudo/gruntExamples.

# 2: CoffeeScript
This plugin is now rewritten in CoffeeScript. Basically, this is still the old code base. However, some functions and loops were already optimized. But there are still some tasks to do…

# 3: SASS
The plugin's stylesheet was converted to a SASS one. Like the CoffeeScript migration, this is not a optimized SASS target.

# 4: Testing / QUnit / PhantomJS
The tests were optimized to use PhantomJS. In addition to that, `grunt` helps to start a mini web server for each test run. With the result that all tests will run in a real environment with no problems of cross origins because of a local file system.

Moreover, an additional `jquery.simulate2` was added supporting a more flexible key event instrument.