Resolved issue #9
Fixed some scopes of event. The first argument 'scope' is always the input field '.as-original' now.
Every change to the selection changes the Placeholder now. If at least one token is selected, the 'placeholder'-attribute is moved to 'disabled-placeholder'. Tf the placeholder is not handled natively by browser ('usePlaceholder' is false), the 'startText'-Option is set as Value of the input field.